Make bracket match styling show up on current line
Fixes #15.

No, this is not the same as 04db131f2281cddc8e92e9ce8b1a9426f27c0b1e.
Some new thing in Atom between 1.19.0 and 1.22.0 caused this to break.

Up the z-index to make the highlights to show up. I _think_ this is the
right fix, but it does feel a little hack-ish.